fixed smoltcp reconnection issue
Previously, the device would not reconnect to MQTT broker after restarting the
program due to smoltcp trying to reuse the previous TCP connection. This was
fixed in https://github.com/smoltcp-rs/smoltcp/pull/367, but this change has
not been released on crates.io yet, so we are pulling in a specific version of
smoltcp until this is released.